fix(cef): address CodeRabbit review findings across CEF engine
- Remove debug stderr prints from main() startup path
- Use strings.HasPrefix for CEF subprocess detection
- Consolidate duplicate resolveEngineType into config.ResolveEngineType()
- Remove vestigial error return from UpdateHistoryTitle
- Guard persistTitleUpdate against empty titles
- Simplify factory adapter returns to direct delegation
- Fix recordBrowserCreateRequest log bug (Msg method capture)
- Make pumpWorkCount atomic to prevent potential races
- Document os.Args mutation thread-safety in engine_init
- Log warning in CreateRelated about unsupported session sharing
- Add error/type logging in pool Prewarm
- Use int64 intermediates in render buffer size calculation
- Distinguish left/right modifier VK codes in input bridge
- Close pendingURI race by holding lock through LoadURL
- Simplify double lock/unlock in webview state update helpers

+++ internal/infrastructure/cef/render_pipeline.go
@@ -159,7 +159,7 @@ func (rp *renderPipeline) handlePaint(buffer unsafe.Pointer, width, height int32
 	startedAt := time.Now()
 	rp.mu.Lock()
 
-	bufSize := int(width) * int(height) * 4
+	bufSize := int(int64(width) * int64(height) * 4)
 	srcSlice := unsafe.Slice((*byte)(buffer), bufSize)
 	copiedBytes := uint64(0)
